refactor : create lotto nums creator interface refactor : lotto can be tested as developer's intention
