+++ src/main/java/domain/LottoNumberCreationStrategy.java
@@ -1,0 +1,7 @@
+package domain;
+
+import java.util.List;
+
+public interface LottoNumberCreationStrategy {
+    public List<Integer> createLottoNums();
+}
