+++ src/test/java/domain/LottoTest.java
@@ -2,13 +2,20 @@ package domain;
 
 import org.junit.Test;
 
+import java.util.Arrays;
+
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class LottoTest {
     @Test
     public void createLottoTicketTest() {
-        Lotto lotto = new Lotto();
-        
-        System.out.println(lotto);
+        Lotto lotto = new Lotto(() -> Arrays.asList(1, 2, 3, 4, 5, 6));
+        assertThat(lotto.toString()).isEqualTo(Arrays.asList(1, 2, 3, 4, 5, 6).toString());
     }
+
+    //당첨 번호를 입력 받으면 해당 로또가 몇개나 일치하는지 리턴
+/*    @Test
+    public void howManyCorrespondTest() {
+
+    }*/
 }
